move tests for consistency
